requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
firecrawl-py==2.15.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-mistralai>=0.0.1
langchain-community>=0.0.10
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.0